refactor(canvas,3d,webgpu,docs): extract gradient and style helpers from CanvasContext to shared utilities and add style properties to CanvasContext3D
- Extract toCanvasGradient, getCanvasGradientBounds, setCanvasFill, and setCanvasStroke to canvas/helpers.ts
- Add style property getters/setters to CanvasContext3D (fill, stroke, opacity, font, shadows, etc.)
- Update CanvasContext to use extracted helper functions for fill and stroke
- Refactor CanvasContext3D face rendering to use layer() and instance
