+++ app/scripts/prepare-playground.mjs
@@ -15,6 +15,7 @@ const PACKAGES = [
     '@ripl/utilities',
     '@ripl/vdom',
     '@ripl/core',
+    '@ripl/dom',
     '@ripl/canvas',
     '@ripl/svg',
     '@ripl/web',
