+++ packages/3d/tsconfig.tsbuildinfo
@@ -1,0 +1,1 @@
+{"root":["../../global.d.ts","./src/camera.ts","./src/constants.ts","./src/context.ts","./src/index.ts","./src/interpolators.ts","./src/shading.ts","./src/core/shape.ts","./src/elements/cone.ts","./src/elements/cube.ts","./src/elements/cylinder.ts","./src/elements/index.ts","./src/elements/plane.ts","./src/elements/sphere.ts","./src/elements/torus.ts","./src/math/index.ts","./src/math/matrix.ts","./src/math/projection.ts","./src/math/vector.ts"],"version":"5.9.3"}
