+++ packages/core/src/core/index.ts
@@ -1,9 +1,9 @@
 export * from './constants';
 export * from './disposer';
 export * from './element';
+export * from './shape';
 export * from './event-bus';
 export * from './factory';
 export * from './group';
-export * from './renderer';
 export * from './scene';
-export * from './shape';
+export * from './renderer';
