--- packages/web/tsup.config.ts
@@ -1,23 +1,0 @@
-import type {
-    Options,
-} from 'tsup';
-
-export default {
-    clean: true,
-    dts: true,
-    sourcemap: true,
-    target: 'es2023',
-    globalName: 'Ripl',
-    outDir: './dist',
-    entry: [
-        './src/index.ts',
-    ],
-    format: [
-        'esm',
-        'cjs',
-        'iife',
-    ],
-    noExternal: [
-        /^@ripl\//,
-    ],
-} as Options;
